Refactor FullTextSearch for testability (and type check the test file!)
