+++ packages/synapse-react-client/src/components/QueryWrapperPlotNav/QueryWrapperPlotNav.tsx
@@ -14,7 +14,7 @@ import { DEFAULT_PAGE_SIZE } from '../../utils/SynapseConstants'
 import { CardConfiguration } from '../CardContainerLogic'
 import { TableQueryDownloadConfirmation } from '../download_list'
 import { SynapseErrorBoundary } from '../error'
-import FullTextSearch from '../FullTextSearch'
+import FullTextSearch from '../FullTextSearch/FullTextSearch'
 import ModalDownload from '../ModalDownload/ModalDownload'
 import { QueryWrapperPlotNavCustomPlotParams } from '../Plot/SynapsePlot'
 import { QueryContextType, useQueryContext } from '../QueryContext'
